Move keyring impl. to keyring, index only exports
